+++ lib/fopost/client.rb
@@ -11,6 +11,7 @@ require 'fopost/resources/inbox'
 require 'fopost/resources/labels'
 require 'fopost/resources/media'
 require 'fopost/resources/posts'
+require 'fopost/resources/blogs'
 require 'fopost/resources/validate'
 require 'fopost/resources/workspaces'
 
@@ -26,7 +27,7 @@ module Fopost
   class Client
     DEFAULT_BASE_URL = HTTP::Client::DEFAULT_BASE_URL
 
-    attr_reader :posts, :accounts, :account_groups, :workspaces, :labels, :ai, :inbox, :ads, :validate, :media
+    attr_reader :posts, :accounts, :account_groups, :workspaces, :labels, :ai, :inbox, :ads, :validate, :media, :blogs
 
     def initialize(api_key: nil, base_url: DEFAULT_BASE_URL, timeout: HTTP::Client::DEFAULT_TIMEOUT,
                    max_retries: HTTP::Client::DEFAULT_MAX_RETRIES, transport: nil, sleeper: nil)
@@ -55,6 +56,7 @@ module Fopost
       @ads = Resources::Ads.new(@http)
       @validate = Resources::Validate.new(@http)
       @media = Resources::Media.new(@http)
+      @blogs = Resources::Blogs.new(@http)
     end
 
     def base_url